Abstract the analytics service
