+++ build.gradle
@@ -2,18 +2,14 @@
 buildscript {
 
     ext {
-        kotlinVersion = '1.5.21'
-        retrofitVersion = '2.9.0'
-        koinVersion = '2.0.1'
-        parcelerVersion = '1.1.13'
-        roomVersion = '2.3.0'
-        navigationVersion = '2.4.0-alpha07'
-        firebaseAnalyticsVersion = '19.0.0'
-        firebaseCrashlyticsVersion = '18.2.1'
-        firebaseMessagingVersion = '20.2.0'
-        firebaseConfigVersion = '19.1.4'
-        adapterDelegatesVersion = '4.2.0'
-        glideVersion = '4.9.0'
+        kotlinVersion = "1.5.21"
+        retrofitVersion = "2.9.0"
+        koinVersion = "2.0.1"
+        roomVersion = "2.3.0"
+        navigationVersion = "2.4.0-alpha07"
+        adapterDelegatesVersion = "4.2.0"
+        glideVersion = "4.9.0"
+        firebaseVersion = "28.3.1"
     }
 
     repositories {
